requirements.txt
flask
dkimpy
pyspf
dnspython
python-dotenv
